% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_CME.R
\name{predictCME}
\alias{predictCME}
\title{predictCME}
\usage{
predictCME(sampler_res, ylist_test, xlist_test, zlist_test, R, S, nom.level)
}
\arguments{
\item{sampler_res}{a list of posterior samples of \eqn{(\beta, \tau^2, \gamma)} returned by the function \code{\link{sampleCME}}.}

\item{ylist_test}{a list of response vectors where the \eqn{i^{th}} component contains \eqn{m_i} observations for the \eqn{i^{th}} subject.}

\item{xlist_test}{a list of fixed effect covariates where the \eqn{i^{th}} component contains \eqn{m_i} observations for the \eqn{i^{th}} subject.}

\item{zlist_test}{a list of random effect covariates where the \eqn{i^{th}} component contains \eqn{m_i} observations for the \eqn{i^{th}} subject.}

\item{R}{the \eqn{k_2 \times q} random projection matrix used to implement the collapsed Gibbs sampler, with entries iid from a Gaussian distribution with mean 0 and variance \eqn{1/k_2}.}

\item{S}{the \eqn{k_1 \times q} random projection matrix used to implement the collapsed Gibbs sampler, with entries iid from a Gaussian distribution with mean 0 and variance \eqn{1/k_1}.}

\item{nom.level}{nominal level for constructing prediction intervals.}
}
\value{
a list containing the following components:
\describe{
\item{ypred}{predicted values of response.}
\item{mspe}{mean square prediction error.}
\item{lower_pi}{lower limits of the prediction intervals.}
\item{upper_pi}{upper limits of the prediction intervals.}
\item{yhat_samples}{posterior samples of the response.}
}
}
\description{
This function performs posterior predictive sampling using the compressed mixed-effects (CME) model.
}
