% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleCME.R
\name{sampleCycle_1}
\alias{sampleCycle_1}
\title{sampleCycle_1}
\usage{
sampleCycle_1(ylist, xlist, zlist, beta, errVar, Gamma, Sigma_gamma, R, S)
}
\arguments{
\item{ylist}{a list of response vectors where the \eqn{i^{th}} component contains \eqn{m_i} observations for the \eqn{i^{th}} subject.}

\item{xlist}{a list of fixed effect covariates where the \eqn{i^{th}} component contains \eqn{m_i \times p} design matrix for the \eqn{i^{th}} subject.}

\item{zlist}{a list of random effect covariates where the \eqn{i^{th}} component contains \eqn{m_i \times q} design matrix for the \eqn{i^{th}} subject.}

\item{beta}{a \eqn{p}-dimensional vector for beta value from previous iteration.}

\item{errVar}{error variance value from previous iteration.}

\item{Gamma}{a \eqn{k_1 \times k_2} matrix for compressed covariance parameter \eqn{\Gamma} from previous iteration.}

\item{Sigma_gamma}{prior covariance matrix of columnwise vectorized \eqn{\Gamma}.}

\item{R}{\eqn{k_2 \times q} random projection matrix with entries iid from a Gaussian distribution with mean 0 and variance \eqn{1/k_2}.}

\item{S}{\eqn{k_1 \times q} random projection matrix with entries iid from a Gaussian distribution with mean 0 and variance \eqn{1/k_1}.}
}
\description{
This function implements the first cycle of the collapsed Gibbs sampler for the CME model by sampling the unobserved compressed random effects and the vectorized compressed covariance parameter from their respective full-conditional distributions.
}
